20240206 Http 요청에 대한 응답 리팩터링

+++ src/main/java/com/miracle/miraclemorningback/service/ComplaintService.java
@@ -3,12 +3,14 @@ package com.miracle.miraclemorningback.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.miracle.miraclemorningback.dto.ComplaintDeleteSuccessResponseDto;
 import com.miracle.miraclemorningback.dto.ComplaintRequestDto;
 import com.miracle.miraclemorningback.dto.ComplaintResponseDto;
+import com.miracle.miraclemorningback.dto.RequestSuccessDto;
 import com.miracle.miraclemorningback.entity.ComplaintEntity;
 import com.miracle.miraclemorningback.repository.ComplaintRepository;
 
@@ -35,28 +37,39 @@ public class ComplaintService {
 
     // 오류 제보
     @Transactional
-    public ComplaintResponseDto complain(ComplaintRequestDto requestDto) {
+    public ResponseEntity<Object> complain(String memberName, ComplaintRequestDto requestDto) {
         ComplaintEntity complaintEntity = ComplaintEntity.builder()
-                .memberName(requestDto.getMemberName())
+                .memberName(memberName)
                 .content(requestDto.getContent())
                 .build();
         complaintRepository.save(complaintEntity);
 
-        return ComplaintResponseDto.builder()
-                .complaintId(complaintEntity.getComplaintId())
-                .memberName(complaintEntity.getMemberName())
-                .content(complaintEntity.getContent())
-                .createdAt(complaintEntity.getCreatedAt())
+        RequestSuccessDto requestSuccessDto = RequestSuccessDto.builder()
+                .success(true)
+                .message("요청이 성공적으로 처리되었습니다.")
                 .build();
+
+        return ResponseEntity.ok().body(requestSuccessDto);
     }
 
     // 제보 삭제
     @Transactional
-    public ComplaintDeleteSuccessResponseDto deleteComplaint(Long complaintId) throws Exception {
-        complaintRepository.findById(complaintId).orElseThrow(
-                () -> new IllegalArgumentException("존재하지 않은 아이디입니다."));
+    public ResponseEntity<Object> deleteComplaint(Long complaintId) {
+        if (!complaintRepository.existsById(complaintId)) {
+            RequestSuccessDto requestSuccessDto = RequestSuccessDto.builder()
+                    .success(false)
+                    .message("해당하는 리소스가 없습니다.")
+                    .build();
 
-        complaintRepository.deleteById(complaintId);
-        return ComplaintDeleteSuccessResponseDto.builder().success(true).build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(requestSuccessDto);
+        } else {
+            complaintRepository.deleteById(complaintId);
+            RequestSuccessDto requestSuccessDto = RequestSuccessDto.builder()
+                    .success(true)
+                    .message("요청이 성공적으로 처리되었습니다.")
+                    .build();
+
+            return ResponseEntity.ok().body(requestSuccessDto);
+        }
     }
 }
